# Fix route matching issue with common prefix