--- a/index.js
+++ b/index.js
@@ -275,7 +275,6 @@ Router.prototype.find = function find (method, path) {
   var pindex = 0
   var params = []
   var i = 0
-  var isStatic = true
 
   while (true) {
     var pathLen = path.length
@@ -309,12 +308,12 @@ Router.prototype.find = function find (method, path) {
     i = pathLen < prefixLen ? pathLen : prefixLen
     while (len < i && path[len] === prefix[len]) len++
 
-    if ((isStatic === true && len > 0) || len === prefixLen) {
+    if (len === prefixLen) {
       path = path.slice(len)
       pathLen = path.length
     }
 
-    var node = currentNode.find(path[0], method)
+    var node = currentNode.find(path, method)
     if (node === null) {
       node = currentNode.parametricBrother
       if (node === null) {
@@ -340,8 +339,6 @@ Router.prototype.find = function find (method, path) {
     if (len !== prefixLen) {
       return getWildcardNode(wildcardNode, method, originalPath, pathLenWildcard)
     }
-
-    isStatic = false
 
     // if exist, save the wildcard child
     if (currentNode.wildcardChild !== null) {