@mcollina 2018-03-12
If we want to speed up string checking, we can also use a XOR as done in here https://github.com/mcollina/bloomrun/blob/master/lib/safeEqual.js.