@delvedor 2018-03-12
A benchmark comparison could be nice!