@nwoltman 2018-03-12
@delvedor Your suggestion did improve the performance a bit. I updated the benchmarks.

@mcollina I tried implementing a custom `startsWith` function that looked similar to that but with `===` and it was slower than the native `string.startsWith()`. Looping across char codes got slowed down in TurboFan, but it should be fast again with V8 66 (reference: https://bugs.chromium.org/p/v8/issues/detail?id=7326&desc=2).